implement convert code without filter
