+++ transcode/transcode_with_filter.h
@@ -1,0 +1,7 @@
+#pragma once
+#ifndef TRANSCODE_WITH_FILTER_H
+#define TRANSCODE_WITH_FILTER_H
+
+int transcode_with_filter();
+
+#endif // !TRANSCODE_WITH_FILTER_H
