+++ transcode/transcode_without_filter.h
@@ -1,0 +1,7 @@
+#pragma once
+#ifndef TRANSCODE_WITHOUT_FILTER_H
+#define TRANSCODE_WITHOUT_FILTER_H
+
+int transcode_without_filter();
+
+#endif // !TRANSCODE_WITHOUT_FILTER_H
